+++ src/Oro/Bundle/CustomerBundle/OroCustomerBundle.php
@@ -3,7 +3,6 @@
 namespace Oro\Bundle\CustomerBundle;
 
 use Oro\Bundle\ApiBundle\DependencyInjection\Compiler\ProcessorBagCompilerPass;
-use Oro\Bundle\CustomerBundle\DependencyInjection\Compiler\ConfigureFrontendHelperPass;
 use Oro\Bundle\CustomerBundle\DependencyInjection\Compiler\DataAuditEntityMappingPass;
 use Oro\Bundle\CustomerBundle\DependencyInjection\Compiler\FrontendApiPass;
 use Oro\Bundle\CustomerBundle\DependencyInjection\Compiler\LoginManagerPass;
@@ -24,7 +23,6 @@ class OroCustomerBundle extends Bundle
         $container->addCompilerPass(new OwnerTreeListenerPass());
         $container->addCompilerPass(new DataAuditEntityMappingPass());
         $container->addCompilerPass(new LoginManagerPass());
-        $container->addCompilerPass(new ConfigureFrontendHelperPass());
 
         if ($container instanceof ExtendedContainerBuilder) {
             $container->addCompilerPass(new FrontendApiPass());