+++ src/Oro/Bundle/FrontendBundle/Resources/public/default/js/app/views/input-widget/responsive-dropdown.js
@@ -212,7 +212,10 @@ const ResponsiveDropdownWidgetView = AbstractInputWidgetView.extend({
 
         const $dropdown = this.$dropdown = $(this.dropdownTemplate(this.getDropdownData()));
 
-        $dropdown.insertBefore(this.$el);
+        $dropdown
+            .data('inputWidget', this)
+            .attr('data-bound-input-widget', this.widgetFunctionName || 'no-name')
+            .insertBefore(this.$el);
         this.$el.detach();
 
         const actions = this.makeDropdownItems();
@@ -272,6 +275,7 @@ const ResponsiveDropdownWidgetView = AbstractInputWidgetView.extend({
         const prepareActions = $actions => {
             return $actions.tooltip('dispose').clone(true)
                 .removeAttr('data-bound-input-widget')
+                .removeAttr('data-bound-component')
                 .removeData('inputWidget');
         };
 
@@ -293,7 +297,7 @@ const ResponsiveDropdownWidgetView = AbstractInputWidgetView.extend({
      * @returns {jQuery.Element}
      */
     getActionsForElement($el) {
-        return $el.find(this.actionsSelector).not(`${this.ignoreActionsIn} ${this.actionsSelector}`);
+        return $el.find(this.actionsSelector).filter(() => !$(this).closest(this.ignoreActionsIn).length);
     },
 
     /**